+++ examples/storybook/stories/Components/ToogleEditorButton.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export const ToggleEditorButton = ({ isTiptap, setIsTiptap }) => {
+  return (
+    <button style={{ fontSize: 18 }} onClick={() => setIsTiptap(!isTiptap)}>
+      {`use ${isTiptap ? 'Draft' : 'Tiptap'}`}
+    </button>
+  );
+};