+++ examples/storybook/stories/Ricos/RicosPublishStory.tsx
@@ -10,6 +10,7 @@ import MobileDetect from 'mobile-detect';
 import ActionButton from '../Components/ActionButton';
 import type { EditorEventsProps } from 'wix-rich-content-editor-common/libs/EditorEventsContext';
 import { withEditorContext } from 'wix-rich-content-editor-common/libs/EditorEventsContext';
+import { ToggleEditorButton } from '../Components/ToogleEditorButton';
 import { POLL_TYPE } from 'ricos-content';
 
 const mobileDetect = new MobileDetect(window.navigator.userAgent);
@@ -40,9 +41,7 @@ const RicosPublishStory: FunctionComponent<EditorEventsProps> = ({ editorEvents 
         </a>
       </h4>
       <Section>
-        <button style={{ fontSize: 18 }} onClick={() => setIsTiptap(!isTiptap)}>
-          {`use ${isTiptap ? 'Draft' : 'Tiptap'}`}
-        </button>
+        <ToggleEditorButton isTiptap={isTiptap} setIsTiptap={setIsTiptap} />
         <RichContentEditorBox>
           <RicosEditor
             experiments={{ tiptapEditor: { enabled: isTiptap } }}