+++ packages/ricos-editor/web/src/utils/bi/publish.ts
@@ -2,7 +2,11 @@ import { Version } from 'wix-rich-content-common';
 import { getEditorContentSummary } from 'wix-rich-content-editor-common';
 import type { DraftContent, BICallbacks } from 'ricos-types';
 
-export const publishBI = (draftContent: DraftContent, onPublish: BICallbacks['onPublish']) => {
+export const publishBI = (
+  draftContent: DraftContent,
+  onPublish: BICallbacks['onPublish'],
+  postId?: string
+) => {
   const { pluginsCount, pluginsDetails } = getEditorContentSummary(draftContent) || {};
-  onPublish?.(undefined, pluginsCount, pluginsDetails, Version.currentVersion, draftContent.ID);
+  onPublish?.(postId, pluginsCount, pluginsDetails, Version.currentVersion, draftContent.ID);
 };